# Do not merge `BenevolentUnionType` with `UnionType`