--- a/src/Type/UnionType.php
+++ b/src/Type/UnionType.php
@@ -70,6 +70,9 @@ class UnionType implements CompoundType
 			if (!($type instanceof UnionType)) {
 				continue;
 			}
+			if ($type instanceof BenevolentUnionType) {
+				continue;
+			}
 			if ($type instanceof TemplateType) {
 				continue;
 			}