--- a/tests/PHPStan/Analyser/nsrt/array-search-php7.php
+++ b/tests/PHPStan/Analyser/nsrt/array-search-php7.php
@@ -12,9 +12,9 @@ class Foo
 	public function mixedAndSubtractedArray($mixed, string $string): void
 	{
 		if (is_array($mixed)) {
-			assertType('int|string|false', array_search('foo', $mixed, true));
-			assertType('int|string|false', array_search('foo', $mixed));
-			assertType('int|string|false', array_search($string, $mixed, true));
+			assertType('(int|string)|false', array_search('foo', $mixed, true));
+			assertType('(int|string)|false', array_search('foo', $mixed));
+			assertType('(int|string)|false', array_search($string, $mixed, true));
 		} else {
 			assertType('mixed~array', $mixed);
 			assertType('null', array_search('foo', $mixed, true));