--- a/tests/PHPStan/Analyser/nsrt/bug-7291.php
+++ b/tests/PHPStan/Analyser/nsrt/bug-7291.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php // lint >= 8.0
+
+declare(strict_types=1);
 
 namespace Bug7291;