@herndlm 2022-12-06
OK close, better than I thought. I guess I deleted too much and am loosing the benevolent template object..