@staabm 2022-12-07
You should measure whether these changes have a perf impact