@VincentLanglet 2022-12-07
Since `A|(B|C)` would be now a possible type, does it require a test to check that a benevolent union
`(A|(B|C)|D)` is simplified to `(A|B|C|D)` ?