@ondrejmirtes 2023-01-08
I'm struggling a bit with this one. What is your thinking behind this PR? I see that it fixes a bunch of issues, but to me it looks like it's just a matter of luck, not an intention (please take no offense :)).

What's a typical situation that this PR improves/fixes?

Also, I think we need some more tests around this. Let's say for example we have `(int|string)` and we want to add some types into it by calling TypeCombinator::union. What's the expected result?

* When we add a `ConstantStringType` like `'1'`.
* When we add a `StringType`.
* When we add ConstantBooleanType like `true`.