@herndlm 2023-01-08
The main intention behind this is that I believe, and I think the linked issues confirm it, that union with a benevolent union should not suddenly get rid of the benevolence. E.g.  https://phpstan.org/r/498cd1d0-4568-4cbf-8b0b-11ed722552e8

Calling `TypeCombinator::union()` will always add a real union and not get rid or extend the benevolence. More tests make sense for sure :)