@herndlm 2023-01-08
Added the test cases for union with `(int|string)` you requested, we get the following

> When we add a `ConstantStringType` like `'1'`.

`(int|string)`


> When we add a `StringType`.

`(int|string)`


>  When we add ConstantBooleanType like `true`.

`(int|string)|true`

that looks OK to me 😊