@ondrejmirtes 2023-01-08
I'd argue that `TypeCombinator::union` of `(int|string)` and `string` should result in non-benevolent union type.

Another idea for a test - if we have `(1|2)` and `string`, it should result in `string` (which it probably will).

Also we can test what happens with `(1|2|3)` and `int<2, 3>`.

Also `(InvalidArgumentException|LengthException|stdClass)` and `LogicException`.