@herndlm 2023-01-08
Ok, added your mentioned cases again, the result is a bit different, I got
- `(1|2)|string`
- `(1|2|3)`
- `(InvalidArgumentException|LengthException|stdClass)|LogicException`

so most likely there's still work needed / it can be improved? I pushed them in this state, but need to think about this another time. Such benevolent cases are still hard for me to think through properly