@VincentLanglet 2023-01-09
> What's a typical situation that this PR improves/fixes?

https://phpstan.org/r/9be20c04-429f-4dc0-b245-9d23628ae2db

Both `$key` and `$bool` are accepted by the method, but the union of both type is not accepted.
The current behavior doesn't feel natural.

> I'd argue that TypeCombinator::union of (int|string) and string should result in non-benevolent union type.

Technically, it should be `(int)|string` which is not really a possible type currently.

Because if it's transformed into `int|string`
```
		$key = current(array_keys($array));
		$this->acceptString($key); // valid
		$this->acceptString($string); // valid
		
		$keyOrString = rand(0,1) ? $key : $string;
		$this->acceptString($keyOrString); // won't be but should still be valid
```