@drupol 2023-01-13
Hello @herndlm,

I don't think this is closing #7423, the output is different for sure, but it's not fixed actually.