@herndlm 2023-01-13
> Sorry, here's the proper link: [phpstan/phpstan#7423](https://github.com/phpstan/phpstan/issues/7423) I also fixed it in my previous message.

yeah and why do you mean this PR does not fix it? the bot message there is unrelated. but in https://github.com/phpstan/phpstan-src/actions/runs/3910108845 I can see no errors for PHP 8+ because the type should be correct now?