@herndlm 2024-10-01
Naming this PR "Do not merge" was a stupid idea, that way it will never get merged... 😂

jokes aside, I try to get this updated and tests added soon. would be great for 2.x I suppose.