@herndlm 2024-10-01
> I'm sorry, I don't see this being merged :/

that's ok, no worries. is there anything I can do to change that or is it mostly based on experience/gut feeling (don't meant that in a bad way)? To me this still makes sense and kind of is the only way IMO to deal with that the linked issues show, also https://github.com/phpstan/phpstan-src/pull/2058#issuecomment-1376493368 and the example there.

but maybe there are edge cases I'm just not seeing yet and already have in mind or are worried about :)