@herndlm 2024-10-01
> I don't want to screw around [phpstan.org/developing-extensions/type-system#type-normalization](https://phpstan.org/developing-extensions/type-system#type-normalization), and keeping benevolent union type separate from non-benevolent union type does that.

sorry, don't want to be too annoying here, but right now they are merged and not merging them keeps them more separate, no? just want to fully understand what might cause issues here.