@herndlm 2024-10-02
what about, yeah I know, I can't leave it alone 😂, the other way around? merging more aggressively: `(int|string)|float` => `(float|int|string)`. my current reasoning: if benevolence is involved type checks are less strict and maybe this should be kept like that if additional types added as in the (outdated) original example of https://phpstan.org/r/498cd1d0-4568-4cbf-8b0b-11ed722552e8

if stricter type checks are needed `checkBenevolentUnionTypes` could be enabled. obviously, I don't know yet what exact impacts this will have, less errors I suppose, so nobody is annoyed. there is of course also https://github.com/phpstan/phpstan/issues/8458#issuecomment-1336402081 which sounds like it would miss more potential runtime errors then I suppose :(